// openapi/models/matter-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Skrib
 * Skrib
 *
 * The version of the OpenAPI document: 2.11.7
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface MatterResponse
 */
export interface MatterResponse {
    /**
     * 
     * @type {object}
     * @memberof MatterResponse
     */
    'events': object;
    /**
     * 
     * @type {number}
     * @memberof MatterResponse
     */
    'id': number;
    /**
     * 
     * @type {string}
     * @memberof MatterResponse
     */
    'title': string;
    /**
     * 
     * @type {string}
     * @memberof MatterResponse
     */
    'description': string;
    /**
     * 
     * @type {number}
     * @memberof MatterResponse
     */
    'tenantid'?: number;
}